requirements.txt
streamlit
streamlit-option-menu
python-dotenv
langchain-core
langchain-community
langchain-experimental
langchain-openai
langchain-text-splitters
langchain-classic
langchain
sentence-transformers
faiss-cpu
networkx
pyvis
py2neo
neo4j
python-docx
pypdf
openai
weaviate-client
rdflib
SPARQLWrapper
